Poly_commitment: move SRS trait implementation where SRS is defined
In future PR, SRS will be moved into evaluation_proof, which will be called ipa.
Step by step.
